+++ .test-infra/jenkins/job_PostCommit_Python_ValidatesRunner_Dataflow.groovy
@@ -20,7 +20,7 @@ import CommonJobProperties as commonJobProperties
 import PostcommitJobBuilder
 
 // This job runs the suite of Python ValidatesRunner tests against the
-// Dataflow runner.
+// Dataflow runner V2.
 PostcommitJobBuilder.postCommitJob('beam_PostCommit_Py_VR_Dataflow', 'Run Python Dataflow ValidatesRunner',
     'Google Cloud Dataflow Runner Python ValidatesRunner Tests', this) {
       description('Runs Python ValidatesRunner suite on the Dataflow runner.')
@@ -36,8 +36,9 @@ PostcommitJobBuilder.postCommitJob('beam_PostCommit_Py_VR_Dataflow', 'Run Python
       steps {
         gradle {
           rootBuildScriptDir(commonJobProperties.checkoutDir)
-          tasks(':sdks:python:test-suites:dataflow:validatesRunnerBatchTests')
-          switches('-Pdisable_runner_v2_until_v2.50')
+          tasks(':sdks:python:test-suites:dataflow:validatesRunnerBatchTestsV2')
+          tasks(':sdks:python:test-suites:dataflow:validatesRunnerStreamingTestsV2')
+          switches('-PuseWheelDistribution')
           commonJobProperties.setGradleSwitches(delegate)
         }
       }