+++ sdks/python/apache_beam/ml/inference/vertex_ai_inference.py
@@ -41,6 +41,22 @@ LOGGER = logging.getLogger("VertexAIModelHandlerJSON")
 # pylint: disable=line-too-long
 
 
+def _retry_on_gcp_client_error(exception):
+  """
+  Retry filter that returns True if a returned HTTP error code is 4xx. This is
+  used to retry remote requests that fail, most notably 429 (TooManyRequests.)
+  This is used for GCP-specific client errors.
+
+  Args:
+    exception: the returned exception encountered during the request/response
+      loop.
+
+  Returns:
+    boolean indication whether or not the exception is a GCP ClientError.
+  """
+  return isinstance(exception, ClientError)
+
+
 class VertexAIModelHandlerJSON(ModelHandler[Any,
                                             PredictionResult,
                                             aiplatform.Endpoint]):
@@ -118,7 +134,8 @@ class VertexAIModelHandlerJSON(ModelHandler[Any,
     ep = self._retrieve_endpoint(self.endpoint_name)
     return ep
 
-  @retry.with_exponential_backoff(num_retries=5)
+  @retry.with_exponential_backoff(
+      num_retries=5, retry_filter=_retry_on_gcp_client_error)
   def get_request(
       self,
       batch: Sequence[Any],
@@ -143,6 +160,9 @@ class VertexAIModelHandlerJSON(ModelHandler[Any,
       raise
     except ClientError as e:
       LOGGER.warning("request failed with error code %i", e.code)
+      raise
+    except Exception as e:
+      LOGGER.error("unexpected exception raised as part of request, got %s", e)
       raise
 
   def run_inference(