+++ sdks/python/apache_beam/runners/dataflow/dataflow_runner.py
@@ -22,26 +22,18 @@ to the Dataflow Service for remote execution by a worker.
 """
 # pytype: skip-file
 
-import base64
 import logging
 import os
 import threading
 import time
-import traceback
 import warnings
 from collections import defaultdict
 from subprocess import DEVNULL
 from typing import TYPE_CHECKING
 from typing import List
-from urllib.parse import quote
-from urllib.parse import quote_from_bytes
-from urllib.parse import unquote_to_bytes
 
 import apache_beam as beam
 from apache_beam import coders
-from apache_beam import error
-from apache_beam.internal import pickler
-from apache_beam.internal.gcp import json_value
 from apache_beam.options.pipeline_options import DebugOptions
 from apache_beam.options.pipeline_options import GoogleCloudOptions
 from apache_beam.options.pipeline_options import SetupOptions
@@ -50,24 +42,13 @@ from apache_beam.options.pipeline_options import TestOptions
 from apache_beam.options.pipeline_options import TypeOptions
 from apache_beam.options.pipeline_options import WorkerOptions
 from apache_beam.portability import common_urns
-from apache_beam.portability.api import beam_runner_api_pb2
-from apache_beam.pvalue import AsSideInput
-from apache_beam.runners.common import DoFnSignature
 from apache_beam.runners.common import group_by_key_input_visitor
-from apache_beam.runners.dataflow.internal import names
 from apache_beam.runners.dataflow.internal.clients import dataflow as dataflow_api
-from apache_beam.runners.dataflow.internal.names import PropertyNames
-from apache_beam.runners.dataflow.internal.names import TransformNames
 from apache_beam.runners.runner import PipelineResult
 from apache_beam.runners.runner import PipelineRunner
 from apache_beam.runners.runner import PipelineState
-from apache_beam.runners.runner import PValueCache
-from apache_beam.transforms import window
-from apache_beam.transforms.display import DisplayData
-from apache_beam.transforms.sideinputs import SIDE_INPUT_PREFIX
 from apache_beam.typehints import typehints
 from apache_beam.utils import processes
-from apache_beam.utils import proto_utils
 from apache_beam.utils.interactive_utils import is_in_notebook
 from apache_beam.utils.plugin import BeamPlugin
 
@@ -103,9 +84,6 @@ class DataflowRunner(PipelineRunner):
 
   # Imported here to avoid circular dependencies.
   # TODO: Remove the apache_beam.pipeline dependency in CreatePTransformOverride
-  from apache_beam.runners.dataflow.ptransform_overrides import CombineValuesPTransformOverride
-  from apache_beam.runners.dataflow.ptransform_overrides import CreatePTransformOverride
-  from apache_beam.runners.dataflow.ptransform_overrides import ReadPTransformOverride
   from apache_beam.runners.dataflow.ptransform_overrides import NativeReadPTransformOverride
 
   # These overrides should be applied before the proto representation of the
@@ -114,19 +92,7 @@ class DataflowRunner(PipelineRunner):
       NativeReadPTransformOverride(),
   ]  # type: List[PTransformOverride]
 
-  # These overrides should be applied after the proto representation of the
-  # graph is created.
-  _NON_PORTABLE_PTRANSFORM_OVERRIDES = [
-      CombineValuesPTransformOverride(),
-      CreatePTransformOverride(),
-      ReadPTransformOverride(),
-  ]  # type: List[PTransformOverride]
-
   def __init__(self, cache=None):
-    # Cache of CloudWorkflowStep protos generated while the runner
-    # "executes" a pipeline.
-    self._cache = cache if cache is not None else PValueCache()
-    self._unique_step_id = 0
     self._default_environment = None
 
   def is_fnapi_compatible(self):
@@ -135,10 +101,6 @@ class DataflowRunner(PipelineRunner):
   def apply(self, transform, input, options):
     _check_and_add_missing_options(options)
     return super().apply(transform, input, options)
-
-  def _get_unique_step_name(self):
-    self._unique_step_id += 1
-    return 's%s' % self._unique_step_id
 
   @staticmethod
   def poll_for_job_completion(
@@ -262,7 +224,7 @@ class DataflowRunner(PipelineRunner):
     return element
 
   @staticmethod
-  def side_input_visitor(is_runner_v2=False, deterministic_key_coders=True):
+  def side_input_visitor(deterministic_key_coders=True):
     # Imported here to avoid circular dependencies.
     # pylint: disable=wrong-import-order, wrong-import-position
     from apache_beam.pipeline import PipelineVisitor
@@ -300,9 +262,8 @@ class DataflowRunner(PipelineRunner):
                   'Unsupported access pattern for %r: %r' %
                   (transform_node.full_label, access_pattern))
             new_side_inputs.append(new_side_input)
-          if is_runner_v2:
-            transform_node.side_inputs = new_side_inputs
-            transform_node.transform.side_inputs = new_side_inputs
+          transform_node.side_inputs = new_side_inputs
+          transform_node.transform.side_inputs = new_side_inputs
 
     return SideInputVisitor()
 
@@ -363,20 +324,12 @@ class DataflowRunner(PipelineRunner):
             not pipeline._options.view_as(
                 TypeOptions).allow_non_deterministic_key_coders))
 
-  def _check_for_unsupported_features_on_non_portable_worker(self, pipeline):
-    pipeline.visit(self.combinefn_visitor())
-
   def run_pipeline(self, pipeline, options, pipeline_proto=None):
     """Remotely executes entire pipeline or parts reachable from node."""
     if _is_runner_v2_disabled(options):
-      debug_options = options.view_as(DebugOptions)
-      if not debug_options.lookup_experiment('disable_runner_v2_until_v2.50'):
-        raise ValueError(
-            'disable_runner_v2 is deprecated in Beam Python ' +
-            beam.version.__version__ +
-            ' and this execution mode will be removed in a future Beam SDK. '
-            'If needed, please use: '
-            '"--experiments=disable_runner_v2_until_v2.50".')
+      raise ValueError(
+          'Disabling Runner V2 no longer supported '
+          'using Beam Python %s.' % beam.version.__version__)
 
     # Label goog-dataflow-notebook if job is started from notebook.
     if is_in_notebook():
@@ -397,26 +350,12 @@ class DataflowRunner(PipelineRunner):
           'Google Cloud Dataflow runner not available, '
           'please install apache_beam[gcp]')
 
-    if pipeline_proto or pipeline.contains_external_transforms:
-      if _is_runner_v2_disabled(options):
-        raise ValueError(
-            'This pipeline contains cross language transforms, '
-            'which requires Runner V2.')
-      if not _is_runner_v2(options):
-        _LOGGER.info(
-            'Automatically enabling Dataflow Runner V2 since the '
-            'pipeline used cross-language transforms.')
-        _add_runner_v2_missing_options(options)
-
-    is_runner_v2 = _is_runner_v2(options)
-    if not is_runner_v2:
-      self._check_for_unsupported_features_on_non_portable_worker(pipeline)
-
     # Convert all side inputs into a form acceptable to Dataflow.
     if pipeline:
+      pipeline.visit(self.combinefn_visitor())
+
       pipeline.visit(
           self.side_input_visitor(
-              _is_runner_v2(options),
               deterministic_key_coders=not options.view_as(
                   TypeOptions).allow_non_deterministic_key_coders))
 
@@ -429,10 +368,6 @@ class DataflowRunner(PipelineRunner):
         warnings.warn(
             "Native sinks no longer implemented; "
             "ignoring use_legacy_bq_sink.")
-
-      from apache_beam.runners.dataflow.ptransform_overrides import GroupIntoBatchesWithShardedKeyPTransformOverride
-      pipeline.replace_all(
-          [GroupIntoBatchesWithShardedKeyPTransformOverride(self, options)])
 
     if pipeline_proto:
       self.proto_pipeline = pipeline_proto
@@ -449,7 +384,7 @@ class DataflowRunner(PipelineRunner):
             self._default_environment.container_image)
       else:
         artifacts = environments.python_sdk_dependencies(options)
-        if artifacts and _is_runner_v2(options):
+        if artifacts:
           _LOGGER.info(
               "Pipeline has additional dependencies to be installed "
               "in SDK worker container, consider using the SDK "
@@ -501,11 +436,6 @@ class DataflowRunner(PipelineRunner):
             known_runner_urns=frozenset(),
             partial=True)
 
-    if not is_runner_v2:
-      # Performing configured PTransform overrides which should not be reflected
-      # in the proto representation of the graph.
-      pipeline.replace_all(DataflowRunner._NON_PORTABLE_PTRANSFORM_OVERRIDES)
-
     # Add setup_options for all the BeamPlugin imports
     setup_options = options.view_as(SetupOptions)
     plugins = BeamPlugin.get_all_plugin_paths()
@@ -522,16 +452,6 @@ class DataflowRunner(PipelineRunner):
           'min_cpu_platform=' + worker_options.min_cpu_platform)
 
     self.job = apiclient.Job(options, self.proto_pipeline)
-
-    # TODO: Consider skipping these for all use_portable_job_submission jobs.
-    if pipeline:
-      # Dataflow Runner v1 requires output type of the Flatten to be the same as
-      # the inputs, hence we enforce that here. Dataflow Runner v2 does not
-      # require this.
-      pipeline.visit(self.flatten_input_visitor())
-
-      # Trigger a traversal of all reachable nodes.
-      self.visit_transforms(pipeline, options)
 
     test_options = options.view_as(TestOptions)
     # If it is a dry run, return without submitting the job.
@@ -557,11 +477,6 @@ class DataflowRunner(PipelineRunner):
     result.metric_results = self._metrics
     return result
 
-  def _get_typehint_based_encoding(self, typehint, window_coder):
-    """Returns an encoding based on a typehint object."""
-    return self._get_cloud_encoding(
-        self._get_coder(typehint, window_coder=window_coder))
-
   @staticmethod
   def _get_coder(typehint, window_coder):
     """Returns a coder based on a typehint object."""
@@ -569,197 +484,6 @@ class DataflowRunner(PipelineRunner):
       return coders.WindowedValueCoder(
           coders.registry.get_coder(typehint), window_coder=window_coder)
     return coders.registry.get_coder(typehint)
-
-  def _get_cloud_encoding(self, coder, unused=None):
-    """Returns an encoding based on a coder object."""
-    if not isinstance(coder, coders.Coder):
-      raise TypeError(
-          'Coder object must inherit from coders.Coder: %s.' % str(coder))
-    return coder.as_cloud_object(self.proto_context.coders)
-
-  def _get_side_input_encoding(self, input_encoding):
-    """Returns an encoding for the output of a view transform.
-
-    Args:
-      input_encoding: encoding of current transform's input. Side inputs need
-        this because the service will check that input and output types match.
-
-    Returns:
-      An encoding that matches the output and input encoding. This is essential
-      for the View transforms introduced to produce side inputs to a ParDo.
-    """
-    return {
-        '@type': 'kind:stream',
-        'component_encodings': [input_encoding],
-        'is_stream_like': {
-            'value': True
-        },
-    }
-
-  def _get_encoded_output_coder(
-      self, transform_node, window_value=True, output_tag=None):
-    """Returns the cloud encoding of the coder for the output of a transform."""
-
-    if output_tag in transform_node.outputs:
-      element_type = transform_node.outputs[output_tag].element_type
-    elif len(transform_node.outputs) == 1:
-      output_tag = DataflowRunner._only_element(transform_node.outputs.keys())
-      # TODO(robertwb): Handle type hints for multi-output transforms.
-      element_type = transform_node.outputs[output_tag].element_type
-
-    else:
-      # TODO(silviuc): Remove this branch (and assert) when typehints are
-      # propagated everywhere. Returning an 'Any' as type hint will trigger
-      # usage of the fallback coder (i.e., cPickler).
-      element_type = typehints.Any
-    if window_value:
-      # All outputs have the same windowing. So getting the coder from an
-      # arbitrary window is fine.
-      output_tag = next(iter(transform_node.outputs.keys()))
-      window_coder = (
-          transform_node.outputs[output_tag].windowing.windowfn.
-          get_window_coder())
-    else:
-      window_coder = None
-    return self._get_typehint_based_encoding(element_type, window_coder)
-
-  def get_pcoll_with_auto_sharding(self):
-    if not hasattr(self, '_pcoll_with_auto_sharding'):
-      return set()
-    return self._pcoll_with_auto_sharding
-
-  def add_pcoll_with_auto_sharding(self, applied_ptransform):
-    if not hasattr(self, '_pcoll_with_auto_sharding'):
-      self.__setattr__('_pcoll_with_auto_sharding', set())
-    output = DataflowRunner._only_element(applied_ptransform.outputs.keys())
-    self._pcoll_with_auto_sharding.add(
-        applied_ptransform.outputs[output]._unique_name())
-
-  def _add_step(self, step_kind, step_label, transform_node, side_tags=()):
-    """Creates a Step object and adds it to the cache."""
-    # Import here to avoid adding the dependency for local running scenarios.
-    # pylint: disable=wrong-import-order, wrong-import-position
-    from apache_beam.runners.dataflow.internal import apiclient
-    step = apiclient.Step(step_kind, self._get_unique_step_name())
-    self.job.proto.steps.append(step.proto)
-    step.add_property(PropertyNames.USER_NAME, step_label)
-    # Cache the node/step association for the main output of the transform node.
-
-    # External transforms may not use 'None' as an output tag.
-    output_tags = ([None] +
-                   list(side_tags) if None in transform_node.outputs.keys() else
-                   list(transform_node.outputs.keys()))
-
-    # We have to cache output for all tags since some transforms may produce
-    # multiple outputs.
-    for output_tag in output_tags:
-      self._cache.cache_output(transform_node, output_tag, step)
-
-    # Finally, we add the display data items to the pipeline step.
-    # If the transform contains no display data then an empty list is added.
-    step.add_property(
-        PropertyNames.DISPLAY_DATA,
-        [
-            item.get_dict()
-            for item in DisplayData.create_from(transform_node.transform).items
-        ])
-
-    if transform_node.resource_hints:
-      step.add_property(
-          PropertyNames.RESOURCE_HINTS,
-          {
-              hint: quote_from_bytes(value)
-              for (hint, value) in transform_node.resource_hints.items()
-          })
-
-    return step
-
-  def _add_singleton_step(
-      self,
-      label,
-      full_label,
-      tag,
-      input_step,
-      windowing_strategy,
-      access_pattern):
-    """Creates a CollectionToSingleton step used to handle ParDo side inputs."""
-    # Import here to avoid adding the dependency for local running scenarios.
-    from apache_beam.runners.dataflow.internal import apiclient
-    step = apiclient.Step(TransformNames.COLLECTION_TO_SINGLETON, label)
-    self.job.proto.steps.append(step.proto)
-    step.add_property(PropertyNames.USER_NAME, full_label)
-    step.add_property(
-        PropertyNames.PARALLEL_INPUT,
-        {
-            '@type': 'OutputReference',
-            PropertyNames.STEP_NAME: input_step.proto.name,
-            PropertyNames.OUTPUT_NAME: input_step.get_output(tag)
-        })
-    step.encoding = self._get_side_input_encoding(input_step.encoding)
-
-    output_info = {
-        PropertyNames.USER_NAME: '%s.%s' % (full_label, PropertyNames.OUTPUT),
-        PropertyNames.ENCODING: step.encoding,
-        PropertyNames.OUTPUT_NAME: PropertyNames.OUT
-    }
-    if common_urns.side_inputs.MULTIMAP.urn == access_pattern:
-      output_info[PropertyNames.USE_INDEXED_FORMAT] = True
-    step.add_property(PropertyNames.OUTPUT_INFO, [output_info])
-
-    step.add_property(
-        PropertyNames.WINDOWING_STRATEGY,
-        self.serialize_windowing_strategy(
-            windowing_strategy, self._default_environment))
-    return step
-
-  def run_Impulse(self, transform_node, options):
-    step = self._add_step(
-        TransformNames.READ, transform_node.full_label, transform_node)
-    step.add_property(PropertyNames.FORMAT, 'impulse')
-    encoded_impulse_element = coders.WindowedValueCoder(
-        coders.BytesCoder(),
-        coders.coders.GlobalWindowCoder()).get_impl().encode_nested(
-            window.GlobalWindows.windowed_value(b''))
-    if _is_runner_v2(options):
-      encoded_impulse_as_str = self.byte_array_to_json_string(
-          encoded_impulse_element)
-    else:
-      encoded_impulse_as_str = base64.b64encode(encoded_impulse_element).decode(
-          'ascii')
-
-    step.add_property(PropertyNames.IMPULSE_ELEMENT, encoded_impulse_as_str)
-
-    step.encoding = self._get_encoded_output_coder(transform_node)
-    step.add_property(
-        PropertyNames.OUTPUT_INFO,
-        [{
-            PropertyNames.USER_NAME: (
-                '%s.%s' % (transform_node.full_label, PropertyNames.OUT)),
-            PropertyNames.ENCODING: step.encoding,
-            PropertyNames.OUTPUT_NAME: PropertyNames.OUT
-        }])
-
-  def run_Flatten(self, transform_node, options):
-    step = self._add_step(
-        TransformNames.FLATTEN, transform_node.full_label, transform_node)
-    inputs = []
-    for one_input in transform_node.inputs:
-      input_step = self._cache.get_pvalue(one_input)
-      inputs.append({
-          '@type': 'OutputReference',
-          PropertyNames.STEP_NAME: input_step.proto.name,
-          PropertyNames.OUTPUT_NAME: input_step.get_output(one_input.tag)
-      })
-    step.add_property(PropertyNames.INPUTS, inputs)
-    step.encoding = self._get_encoded_output_coder(transform_node)
-    step.add_property(
-        PropertyNames.OUTPUT_INFO,
-        [{
-            PropertyNames.USER_NAME: (
-                '%s.%s' % (transform_node.full_label, PropertyNames.OUT)),
-            PropertyNames.ENCODING: step.encoding,
-            PropertyNames.OUTPUT_NAME: PropertyNames.OUT
-        }])
 
   # TODO(srohde): Remove this after internal usages have been removed.
   def apply_GroupByKey(self, transform, pcoll, options):
@@ -783,512 +507,6 @@ class DataflowRunner(PipelineRunner):
     # TODO(robertwb): Update the coder itself if it changed.
     coders.registry.verify_deterministic(
         coder.key_coder(), 'GroupByKey operation "%s"' % transform.label)
-
-  def run_GroupByKey(self, transform_node, options):
-    input_tag = transform_node.inputs[0].tag
-    input_step = self._cache.get_pvalue(transform_node.inputs[0])
-
-    # Verify that the GBK's parent has a KV coder.
-    self._verify_gbk_coders(transform_node.transform, transform_node.inputs[0])
-
-    step = self._add_step(
-        TransformNames.GROUP, transform_node.full_label, transform_node)
-    step.add_property(
-        PropertyNames.PARALLEL_INPUT,
-        {
-            '@type': 'OutputReference',
-            PropertyNames.STEP_NAME: input_step.proto.name,
-            PropertyNames.OUTPUT_NAME: input_step.get_output(input_tag)
-        })
-    step.encoding = self._get_encoded_output_coder(transform_node)
-    step.add_property(
-        PropertyNames.OUTPUT_INFO,
-        [{
-            PropertyNames.USER_NAME: (
-                '%s.%s' % (transform_node.full_label, PropertyNames.OUT)),
-            PropertyNames.ENCODING: step.encoding,
-            PropertyNames.OUTPUT_NAME: PropertyNames.OUT
-        }])
-    windowing = transform_node.transform.get_windowing(transform_node.inputs)
-    step.add_property(
-        PropertyNames.SERIALIZED_FN,
-        self.serialize_windowing_strategy(windowing, self._default_environment))
-
-  def run_ExternalTransform(self, transform_node, options):
-    # Adds a dummy step to the Dataflow job description so that inputs and
-    # outputs are mapped correctly in the presence of external transforms.
-    #
-    # Note that Dataflow Python multi-language pipelines use Portable Job
-    # Submission by default, hence this step and rest of the Dataflow step
-    # definitions defined here are not used at Dataflow service but we have to
-    # maintain the mapping correctly till we can fully drop the Dataflow step
-    # definitions from the SDK.
-
-    # AppliedTransform node outputs have to be updated to correctly map the
-    # outputs for external transforms.
-    transform_node.outputs = ({
-        output.tag: output
-        for output in transform_node.outputs.values()
-    })
-
-    self.run_Impulse(transform_node, options)
-
-  def run_ParDo(self, transform_node, options):
-    transform = transform_node.transform
-    input_tag = transform_node.inputs[0].tag
-    input_step = self._cache.get_pvalue(transform_node.inputs[0])
-
-    # Attach side inputs.
-    si_dict = {}
-    si_labels = {}
-    full_label_counts = defaultdict(int)
-    lookup_label = lambda side_pval: si_labels[side_pval]
-    named_inputs = transform_node.named_inputs()
-    label_renames = {}
-    for ix, side_pval in enumerate(transform_node.side_inputs):
-      assert isinstance(side_pval, AsSideInput)
-      step_name = 'SideInput-' + self._get_unique_step_name()
-      si_label = ((SIDE_INPUT_PREFIX + '%d-%s') %
-                  (ix, transform_node.full_label))
-      old_label = (SIDE_INPUT_PREFIX + '%d') % ix
-
-      label_renames[old_label] = si_label
-
-      assert old_label in named_inputs
-      pcollection_label = '%s.%s' % (
-          side_pval.pvalue.producer.full_label.split('/')[-1],
-          side_pval.pvalue.tag if side_pval.pvalue.tag else 'out')
-      si_full_label = '%s/%s(%s.%s)' % (
-          transform_node.full_label,
-          side_pval.__class__.__name__,
-          pcollection_label,
-          full_label_counts[pcollection_label])
-
-      # Count the number of times the same PCollection is a side input
-      # to the same ParDo.
-      full_label_counts[pcollection_label] += 1
-
-      self._add_singleton_step(
-          step_name,
-          si_full_label,
-          side_pval.pvalue.tag,
-          self._cache.get_pvalue(side_pval.pvalue),
-          side_pval.pvalue.windowing,
-          side_pval._side_input_data().access_pattern)
-      si_dict[si_label] = {
-          '@type': 'OutputReference',
-          PropertyNames.STEP_NAME: step_name,
-          PropertyNames.OUTPUT_NAME: PropertyNames.OUT
-      }
-      si_labels[side_pval] = si_label
-
-    # Now create the step for the ParDo transform being handled.
-    transform_name = transform_node.full_label.rsplit('/', 1)[-1]
-    step = self._add_step(
-        TransformNames.DO,
-        transform_node.full_label +
-        ('/{}'.format(transform_name) if transform_node.side_inputs else ''),
-        transform_node,
-        transform_node.transform.output_tags)
-    transform_proto = self.proto_context.transforms.get_proto(transform_node)
-    transform_id = self.proto_context.transforms.get_id(transform_node)
-    is_runner_v2 = _is_runner_v2(options)
-    # Patch side input ids to be unique across a given pipeline.
-    if (label_renames and
-        transform_proto.spec.urn == common_urns.primitives.PAR_DO.urn):
-      # Patch PTransform proto.
-      for old, new in label_renames.items():
-        transform_proto.inputs[new] = transform_proto.inputs[old]
-        del transform_proto.inputs[old]
-
-      # Patch ParDo proto.
-      proto_type, _ = beam.PTransform._known_urns[transform_proto.spec.urn]
-      proto = proto_utils.parse_Bytes(transform_proto.spec.payload, proto_type)
-      for old, new in label_renames.items():
-        proto.side_inputs[new].CopyFrom(proto.side_inputs[old])
-        del proto.side_inputs[old]
-      transform_proto.spec.payload = proto.SerializeToString()
-      # We need to update the pipeline proto.
-      del self.proto_pipeline.components.transforms[transform_id]
-      (
-          self.proto_pipeline.components.transforms[transform_id].CopyFrom(
-              transform_proto))
-    # The data transmitted in SERIALIZED_FN is different depending on whether
-    # this is a runner v2 pipeline or not.
-    if is_runner_v2:
-      serialized_data = transform_id
-    else:
-      serialized_data = pickler.dumps(
-          self._pardo_fn_data(transform_node, lookup_label))
-    step.add_property(PropertyNames.SERIALIZED_FN, serialized_data)
-    # TODO(BEAM-8882): Enable once dataflow service doesn't reject this.
-    # step.add_property(PropertyNames.PIPELINE_PROTO_TRANSFORM_ID, transform_id)
-    step.add_property(
-        PropertyNames.PARALLEL_INPUT,
-        {
-            '@type': 'OutputReference',
-            PropertyNames.STEP_NAME: input_step.proto.name,
-            PropertyNames.OUTPUT_NAME: input_step.get_output(input_tag)
-        })
-    # Add side inputs if any.
-    step.add_property(PropertyNames.NON_PARALLEL_INPUTS, si_dict)
-
-    # Generate description for the outputs. The output names
-    # will be 'None' for main output and '<tag>' for a tagged output.
-    outputs = []
-
-    all_output_tags = list(transform_proto.outputs.keys())
-
-    # Some external transforms require output tags to not be modified.
-    # So we randomly select one of the output tags as the main output and
-    # leave others as side outputs. Transform execution should not change
-    # dependending on which output tag we choose as the main output here.
-    # Also, some SDKs do not work correctly if output tags are modified. So for
-    # external transforms, we leave tags unmodified.
-    #
-    # Python SDK uses 'None' as the tag of the main output.
-    main_output_tag = 'None'
-
-    step.encoding = self._get_encoded_output_coder(
-        transform_node, output_tag=main_output_tag)
-
-    side_output_tags = set(all_output_tags).difference({main_output_tag})
-
-    # Add the main output to the description.
-    outputs.append({
-        PropertyNames.USER_NAME: (
-            '%s.%s' % (transform_node.full_label, PropertyNames.OUT)),
-        PropertyNames.ENCODING: step.encoding,
-        PropertyNames.OUTPUT_NAME: main_output_tag
-    })
-    for side_tag in side_output_tags:
-      # The assumption here is that all outputs will have the same typehint
-      # and coder as the main output. This is certainly the case right now
-      # but conceivably it could change in the future.
-      encoding = self._get_encoded_output_coder(
-          transform_node, output_tag=side_tag)
-      outputs.append({
-          PropertyNames.USER_NAME: (
-              '%s.%s' % (transform_node.full_label, side_tag)),
-          PropertyNames.ENCODING: encoding,
-          PropertyNames.OUTPUT_NAME: side_tag
-      })
-
-    step.add_property(PropertyNames.OUTPUT_INFO, outputs)
-
-    # Add the restriction encoding if we are a splittable DoFn
-    restriction_coder = transform.get_restriction_coder()
-    if restriction_coder:
-      step.add_property(
-          PropertyNames.RESTRICTION_ENCODING,
-          self._get_cloud_encoding(restriction_coder))
-
-    if options.view_as(StandardOptions).streaming:
-      is_stateful_dofn = (DoFnSignature(transform.dofn).is_stateful_dofn())
-      if is_stateful_dofn:
-        step.add_property(PropertyNames.USES_KEYED_STATE, 'true')
-
-        # Also checks whether the step allows shardable keyed states.
-        # TODO(BEAM-11360): remove this when migrated to portable job
-        #  submission since we only consider supporting the property in runner
-        #  v2.
-        for pcoll in transform_node.outputs.values():
-          if pcoll._unique_name() in self.get_pcoll_with_auto_sharding():
-            step.add_property(PropertyNames.ALLOWS_SHARDABLE_STATE, 'true')
-            # Currently we only allow auto-sharding to be enabled through the
-            # GroupIntoBatches transform. So we also add the following property
-            # which GroupIntoBatchesDoFn has, to allow the backend to perform
-            # graph optimization.
-            step.add_property(PropertyNames.PRESERVES_KEYS, 'true')
-            break
-
-  @staticmethod
-  def _pardo_fn_data(transform_node, get_label):
-    transform = transform_node.transform
-    si_tags_and_types = [  # pylint: disable=protected-access
-        (get_label(side_pval), side_pval.__class__, side_pval._view_options())
-        for side_pval in transform_node.side_inputs]
-    return (
-        transform.fn,
-        transform.args,
-        transform.kwargs,
-        si_tags_and_types,
-        transform_node.inputs[0].windowing)
-
-  def run_CombineValuesReplacement(self, transform_node, options):
-    transform = transform_node.transform.transform
-    input_tag = transform_node.inputs[0].tag
-    input_step = self._cache.get_pvalue(transform_node.inputs[0])
-    step = self._add_step(
-        TransformNames.COMBINE, transform_node.full_label, transform_node)
-    transform_id = self.proto_context.transforms.get_id(transform_node.parent)
-
-    # The data transmitted in SERIALIZED_FN is different depending on whether
-    # this is a runner v2 pipeline or not.
-    if _is_runner_v2(options):
-      # Fnapi pipelines send the transform ID of the CombineValues transform's
-      # parent composite because Dataflow expects the ID of a CombinePerKey
-      # transform.
-      serialized_data = transform_id
-    else:
-      # Combiner functions do not take deferred side-inputs (i.e. PValues) and
-      # therefore the code to handle extra args/kwargs is simpler than for the
-      # DoFn's of the ParDo transform. In the last, empty argument is where
-      # side inputs information would go.
-      serialized_data = pickler.dumps(
-          (transform.fn, transform.args, transform.kwargs, ()))
-    step.add_property(PropertyNames.SERIALIZED_FN, serialized_data)
-    # TODO(BEAM-8882): Enable once dataflow service doesn't reject this.
-    # step.add_property(PropertyNames.PIPELINE_PROTO_TRANSFORM_ID, transform_id)
-    step.add_property(
-        PropertyNames.PARALLEL_INPUT,
-        {
-            '@type': 'OutputReference',
-            PropertyNames.STEP_NAME: input_step.proto.name,
-            PropertyNames.OUTPUT_NAME: input_step.get_output(input_tag)
-        })
-    # Note that the accumulator must not have a WindowedValue encoding, while
-    # the output of this step does in fact have a WindowedValue encoding.
-    accumulator_encoding = self._get_cloud_encoding(
-        transform.fn.get_accumulator_coder())
-    output_encoding = self._get_encoded_output_coder(transform_node)
-
-    step.encoding = output_encoding
-    step.add_property(PropertyNames.ENCODING, accumulator_encoding)
-    # Generate description for main output 'out.'
-    outputs = []
-    # Add the main output to the description.
-    outputs.append({
-        PropertyNames.USER_NAME: (
-            '%s.%s' % (transform_node.full_label, PropertyNames.OUT)),
-        PropertyNames.ENCODING: step.encoding,
-        PropertyNames.OUTPUT_NAME: PropertyNames.OUT
-    })
-    step.add_property(PropertyNames.OUTPUT_INFO, outputs)
-
-  def run_Read(self, transform_node, options):
-    transform = transform_node.transform
-    step = self._add_step(
-        TransformNames.READ, transform_node.full_label, transform_node)
-    # TODO(mairbek): refactor if-else tree to use registerable functions.
-    # Initialize the source specific properties.
-
-    standard_options = options.view_as(StandardOptions)
-    if not hasattr(transform.source, 'format'):
-      # If a format is not set, we assume the source to be a custom source.
-      source_dict = {}
-
-      source_dict['spec'] = {
-          '@type': names.SOURCE_TYPE,
-          names.SERIALIZED_SOURCE_KEY: pickler.dumps(transform.source)
-      }
-
-      try:
-        source_dict['metadata'] = {
-            'estimated_size_bytes': json_value.get_typed_value_descriptor(
-                transform.source.estimate_size())
-        }
-      except error.RuntimeValueProviderError:
-        # Size estimation is best effort, and this error is by value provider.
-        _LOGGER.info(
-            'Could not estimate size of source %r due to ' + \
-            'RuntimeValueProviderError', transform.source)
-      except Exception:  # pylint: disable=broad-except
-        # Size estimation is best effort. So we log the error and continue.
-        _LOGGER.info(
-            'Could not estimate size of source %r due to an exception: %s',
-            transform.source,
-            traceback.format_exc())
-
-      step.add_property(PropertyNames.SOURCE_STEP_INPUT, source_dict)
-    elif transform.source.format == 'pubsub':
-      if not standard_options.streaming:
-        raise ValueError(
-            'Cloud Pub/Sub is currently available for use '
-            'only in streaming pipelines.')
-      # Only one of topic or subscription should be set.
-      if transform.source.full_subscription:
-        step.add_property(
-            PropertyNames.PUBSUB_SUBSCRIPTION,
-            transform.source.full_subscription)
-      elif transform.source.full_topic:
-        step.add_property(
-            PropertyNames.PUBSUB_TOPIC, transform.source.full_topic)
-      if transform.source.id_label:
-        step.add_property(
-            PropertyNames.PUBSUB_ID_LABEL, transform.source.id_label)
-      if transform.source.with_attributes:
-        # Setting this property signals Dataflow runner to return full
-        # PubsubMessages instead of just the data part of the payload.
-        step.add_property(PropertyNames.PUBSUB_SERIALIZED_ATTRIBUTES_FN, '')
-
-      if transform.source.timestamp_attribute is not None:
-        step.add_property(
-            PropertyNames.PUBSUB_TIMESTAMP_ATTRIBUTE,
-            transform.source.timestamp_attribute)
-    else:
-      raise ValueError(
-          'Source %r has unexpected format %s.' %
-          (transform.source, transform.source.format))
-
-    if not hasattr(transform.source, 'format'):
-      step.add_property(PropertyNames.FORMAT, names.SOURCE_FORMAT)
-    else:
-      step.add_property(PropertyNames.FORMAT, transform.source.format)
-
-    # Wrap coder in WindowedValueCoder: this is necessary as the encoding of a
-    # step should be the type of value outputted by each step.  Read steps
-    # automatically wrap output values in a WindowedValue wrapper, if necessary.
-    # This is also necessary for proper encoding for size estimation.
-    # Using a GlobalWindowCoder as a place holder instead of the default
-    # PickleCoder because GlobalWindowCoder is known coder.
-    # TODO(robertwb): Query the collection for the windowfn to extract the
-    # correct coder.
-    coder = coders.WindowedValueCoder(
-        coders.registry.get_coder(transform_node.outputs[None].element_type),
-        coders.coders.GlobalWindowCoder())
-
-    step.encoding = self._get_cloud_encoding(coder)
-    step.add_property(
-        PropertyNames.OUTPUT_INFO,
-        [{
-            PropertyNames.USER_NAME: (
-                '%s.%s' % (transform_node.full_label, PropertyNames.OUT)),
-            PropertyNames.ENCODING: step.encoding,
-            PropertyNames.OUTPUT_NAME: PropertyNames.OUT
-        }])
-
-  def run__NativeWrite(self, transform_node, options):
-    transform = transform_node.transform
-    input_tag = transform_node.inputs[0].tag
-    input_step = self._cache.get_pvalue(transform_node.inputs[0])
-    step = self._add_step(
-        TransformNames.WRITE, transform_node.full_label, transform_node)
-    # TODO(mairbek): refactor if-else tree to use registerable functions.
-    # Initialize the sink specific properties.
-    if transform.sink.format == 'pubsub':
-      standard_options = options.view_as(StandardOptions)
-      if not standard_options.streaming:
-        raise ValueError(
-            'Cloud Pub/Sub is currently available for use '
-            'only in streaming pipelines.')
-      step.add_property(PropertyNames.PUBSUB_TOPIC, transform.sink.full_topic)
-      if transform.sink.id_label:
-        step.add_property(
-            PropertyNames.PUBSUB_ID_LABEL, transform.sink.id_label)
-      # Setting this property signals Dataflow runner that the PCollection
-      # contains PubsubMessage objects instead of just raw data.
-      step.add_property(PropertyNames.PUBSUB_SERIALIZED_ATTRIBUTES_FN, '')
-      if transform.sink.timestamp_attribute is not None:
-        step.add_property(
-            PropertyNames.PUBSUB_TIMESTAMP_ATTRIBUTE,
-            transform.sink.timestamp_attribute)
-    else:
-      raise ValueError(
-          'Sink %r has unexpected format %s.' %
-          (transform.sink, transform.sink.format))
-    step.add_property(PropertyNames.FORMAT, transform.sink.format)
-
-    # Wrap coder in WindowedValueCoder: this is necessary for proper encoding
-    # for size estimation. Using a GlobalWindowCoder as a place holder instead
-    # of the default PickleCoder because GlobalWindowCoder is known coder.
-    # TODO(robertwb): Query the collection for the windowfn to extract the
-    # correct coder.
-    coder = coders.WindowedValueCoder(
-        transform.sink.coder, coders.coders.GlobalWindowCoder())
-    step.encoding = self._get_cloud_encoding(coder)
-    step.add_property(PropertyNames.ENCODING, step.encoding)
-    step.add_property(
-        PropertyNames.PARALLEL_INPUT,
-        {
-            '@type': 'OutputReference',
-            PropertyNames.STEP_NAME: input_step.proto.name,
-            PropertyNames.OUTPUT_NAME: input_step.get_output(input_tag)
-        })
-
-  def run_TestStream(self, transform_node, options):
-    from apache_beam.testing.test_stream import ElementEvent
-    from apache_beam.testing.test_stream import ProcessingTimeEvent
-    from apache_beam.testing.test_stream import WatermarkEvent
-    standard_options = options.view_as(StandardOptions)
-    if not standard_options.streaming:
-      raise ValueError(
-          'TestStream is currently available for use '
-          'only in streaming pipelines.')
-
-    transform = transform_node.transform
-    step = self._add_step(
-        TransformNames.READ, transform_node.full_label, transform_node)
-    step.add_property(
-        PropertyNames.SERIALIZED_FN,
-        self.proto_context.transforms.get_id(transform_node))
-    step.add_property(PropertyNames.FORMAT, 'test_stream')
-    test_stream_payload = beam_runner_api_pb2.TestStreamPayload()
-    # TestStream source doesn't do any decoding of elements,
-    # so we won't set test_stream_payload.coder_id.
-    output_coder = transform._infer_output_coder()  # pylint: disable=protected-access
-    for event in transform._events:
-      new_event = test_stream_payload.events.add()
-      if isinstance(event, ElementEvent):
-        for tv in event.timestamped_values:
-          element = new_event.element_event.elements.add()
-          element.encoded_element = output_coder.encode(tv.value)
-          element.timestamp = tv.timestamp.micros
-      elif isinstance(event, ProcessingTimeEvent):
-        new_event.processing_time_event.advance_duration = (
-            event.advance_by.micros)
-      elif isinstance(event, WatermarkEvent):
-        new_event.watermark_event.new_watermark = event.new_watermark.micros
-    serialized_payload = self.byte_array_to_json_string(
-        test_stream_payload.SerializeToString())
-    step.add_property(PropertyNames.SERIALIZED_TEST_STREAM, serialized_payload)
-
-    step.encoding = self._get_encoded_output_coder(transform_node)
-    step.add_property(
-        PropertyNames.OUTPUT_INFO,
-        [{
-            PropertyNames.USER_NAME: (
-                '%s.%s' % (transform_node.full_label, PropertyNames.OUT)),
-            PropertyNames.ENCODING: step.encoding,
-            PropertyNames.OUTPUT_NAME: PropertyNames.OUT
-        }])
-
-  # We must mark this method as not a test or else its name is a matcher for
-  # nosetest tests.
-  run_TestStream.__test__ = False  # type: ignore[attr-defined]
-
-  @classmethod
-  def serialize_windowing_strategy(cls, windowing, default_environment):
-    from apache_beam.runners import pipeline_context
-    context = pipeline_context.PipelineContext(
-        default_environment=default_environment)
-    windowing_proto = windowing.to_runner_api(context)
-    return cls.byte_array_to_json_string(
-        beam_runner_api_pb2.MessageWithComponents(
-            components=context.to_runner_api(),
-            windowing_strategy=windowing_proto).SerializeToString())
-
-  @classmethod
-  def deserialize_windowing_strategy(cls, serialized_data):
-    # Imported here to avoid circular dependencies.
-    # pylint: disable=wrong-import-order, wrong-import-position
-    from apache_beam.runners import pipeline_context
-    from apache_beam.transforms.core import Windowing
-    proto = beam_runner_api_pb2.MessageWithComponents()
-    proto.ParseFromString(cls.json_string_to_byte_array(serialized_data))
-    return Windowing.from_runner_api(
-        proto.windowing_strategy,
-        pipeline_context.PipelineContext(proto.components))
-
-  @staticmethod
-  def byte_array_to_json_string(raw_bytes):
-    """Implements org.apache.beam.sdk.util.StringUtils.byteArrayToJsonString."""
-    return quote(raw_bytes)
-
-  @staticmethod
-  def json_string_to_byte_array(encoded_string):
-    """Implements org.apache.beam.sdk.util.StringUtils.jsonStringToByteArray."""
-    return unquote_to_bytes(encoded_string)
 
   def get_default_gcp_region(self):
     """Get a default value for Google Cloud region according to
@@ -1348,6 +566,8 @@ def _check_and_add_missing_options(options):
   options.view_as(
       GoogleCloudOptions).dataflow_service_options = dataflow_service_options
 
+  _add_runner_v2_missing_options(options)
+
   # Ensure that prime is specified as an experiment if specified as a dataflow
   # service option
   if 'enable_prime' in dataflow_service_options:
@@ -1359,11 +579,6 @@ def _check_and_add_missing_options(options):
   # Runner v2 only supports using streaming engine (aka windmill service)
   if options.view_as(StandardOptions).streaming:
     google_cloud_options = options.view_as(GoogleCloudOptions)
-    if _is_runner_v2_disabled(options):
-      raise ValueError(
-          'Disabling Runner V2 no longer supported for streaming pipeline '
-          'using Beam Python %s.' % beam.version.__version__)
-
     if (not google_cloud_options.enable_streaming_engine and
         (debug_options.lookup_experiment("enable_windmill_service") or
          debug_options.lookup_experiment("enable_streaming_engine"))):
@@ -1380,29 +595,6 @@ def _check_and_add_missing_options(options):
     google_cloud_options.enable_streaming_engine = True
     debug_options.add_experiment("enable_streaming_engine")
     debug_options.add_experiment("enable_windmill_service")
-    _add_runner_v2_missing_options(debug_options)
-  elif (debug_options.lookup_experiment('enable_prime') or
-        debug_options.lookup_experiment('beam_fn_api') or
-        debug_options.lookup_experiment('use_unified_worker') or
-        debug_options.lookup_experiment('use_runner_v2') or
-        debug_options.lookup_experiment('use_portable_job_submission')):
-    if _is_runner_v2_disabled(options):
-      raise ValueError(
-          """Runner V2 both disabled and enabled: at least one of
-          ['enable_prime', 'beam_fn_api', 'use_unified_worker', 'use_runner_v2',
-          'use_portable_job_submission'] is set and also one of
-          ['disable_runner_v2', 'disable_runner_v2_until_2023',
-          'disable_prime_runner_v2'] is set.""")
-    _add_runner_v2_missing_options(debug_options)
-
-
-def _is_runner_v2(options):
-  # Type: (PipelineOptions) -> bool
-
-  """Returns true if runner v2 is enabled."""
-  _check_and_add_missing_options(options)
-  return options.view_as(DebugOptions).lookup_experiment(
-      'use_runner_v2', default=False)
 
 
 def _is_runner_v2_disabled(options):